requirements.txt
Django>=4.2
pillow>=10.4.0
django-phonenumber-field>=8.0.0
phonenumbers>=8.13.39
python-dotenv>=1.0.1
environs>=11.0.0
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.3.1
djoser>=2.2.3
social-auth-app-django>=5.4.2
djangorestframework-jwt
